Updated: Inserindo conexão com banco de dados PostgreSQL hospedado em container docker

+++ src/main/resources/db/migration/V1__Create_Table_Users.sql
@@ -1,7 +1,7 @@
 CREATE TABLE tb_users(
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    weight DOUBLE NOT NULL,
+    weight DOUBLE PRECISION NOT NULL,
     email VARCHAR(255) NOT NULL,
     password VARCHAR(255) NOT NULL
 );
